add raw transcript for episode 339
For personal research, I found it useful to make a transcript for episode 339 ([TEEs with Andrew Miller](https://zeroknowledge.fm/podcast/339/)) using a local instance of Whisper. The transctipt is not edited, has no timestamps, and does not distinguish between speakers. I thought that maybe it could still be useful (it certainly was for me).